requirements.txt
colorama
requests
pyotp
python-dotenv
websocket-client
win11toast; platform_system == "Windows"
